backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pymongo==4.6.1
motor==3.3.2
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
pinecone-client==3.0.0
pypdf==4.0.1
python-docx==1.1.0
ffmpeg-python==0.2.0
pydub==0.25.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
